+++ engines/cel/engine.go
@@ -3,8 +3,8 @@ package tempo_cel
 import (
 	"io"
 
-	"github.com/tniedbala/tempe-go/pkg"
-	"github.com/tniedbala/tempe-go/pkg/api"
+	"github.com/tniedbala/tempe-go/tempe"
+	"github.com/tniedbala/tempe-go/tempe/api"
 )
 
 type CelEngine struct {
