+++ engines/starlark/value.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"iter"
 
-	"github.com/tniedbala/tempe-go/pkg/api"
+	"github.com/tniedbala/tempe-go/tempe/api"
 	"go.starlark.net/starlark"
 )
