+++ scripts/build-go.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 GRAMMAR_DIR=$WORKSPACE/grammar
-PACKAGE_DIR=$WORKSPACE/pkg/parser/base
+PACKAGE_DIR=$WORKSPACE/tempe/parser/base
 
 rm -rf $PACKAGE_DIR
 antlr -o $PACKAGE_DIR -Dlanguage=Go -package base -visitor -no-listener $GRAMMAR_DIR/TemplateLexer.g4
