+++ tempe/go_engine.go
@@ -3,7 +3,7 @@ package tempe
 import (
 	"io"
 
-	"github.com/tniedbala/tempe-go/pkg/api"
+	"github.com/tniedbala/tempe-go/tempe/api"
 )
 
 type GoEngine struct {}
