+++ tempe/go_value.go
@@ -5,7 +5,7 @@ import (
 	"iter"
 	"reflect"
 
-	"github.com/tniedbala/tempe-go/pkg/api"
+	"github.com/tniedbala/tempe-go/tempe/api"
 )
 
 type GoValue struct {
