+++ tempe/parser/parse_tree.go
@@ -3,7 +3,7 @@ package parser
 import (
 	"iter"
 
-	"github.com/tniedbala/tempe-go/pkg/api"
+	"github.com/tniedbala/tempe-go/tempe/api"
 )
 
 type ParseTree struct {
