+++ tempe/parser/template_nodes_collection.go
@@ -4,7 +4,7 @@ import (
 	"io"
 	"iter"
 
-	"github.com/tniedbala/tempe-go/pkg/api"
+	"github.com/tniedbala/tempe-go/tempe/api"
 )
 
 type TemplateNodesCollection struct {
