+++ tempe/parser/template_visitor.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 
 	"github.com/antlr4-go/antlr/v4"
-	"github.com/tniedbala/tempe-go/pkg/api"
-	"github.com/tniedbala/tempe-go/pkg/parser/base"
+	"github.com/tniedbala/tempe-go/tempe/api"
+	"github.com/tniedbala/tempe-go/tempe/parser/base"
 )
 
 type TemplateVisitor struct {
