+++ tempe/template.go
@@ -6,8 +6,8 @@ import (
 	"io"
 	"os"
 
-	"github.com/tniedbala/tempe-go/pkg/api"
-	"github.com/tniedbala/tempe-go/pkg/parser"
+	"github.com/tniedbala/tempe-go/tempe/api"
+	"github.com/tniedbala/tempe-go/tempe/parser"
 )
 
 type Template struct {
